+++ components/updater/card/src/main/java/com/flipperdevices/updater/card/model/FlipperState.kt
@@ -1,9 +1,11 @@
 package com.flipperdevices.updater.card.model
 
-enum class FlipperState {
-    NOT_READY,
-    UPDATING,
-    COMPLETE,
-    FAILED,
-    READY
+import com.flipperdevices.updater.model.FirmwareVersion
+
+sealed class FlipperState {
+    object NotReady : FlipperState()
+    object Updating : FlipperState()
+    class Complete(val version: FirmwareVersion?) : FlipperState()
+    class Failed(val version: FirmwareVersion?) : FlipperState()
+    object Ready : FlipperState()
 }