+++ components/updater/screen/src/main/java/com/flipperdevices/updater/screen/model/UpdaterScreenState.kt
@@ -19,6 +19,10 @@ sealed class UpdaterScreenState(
         val percent: Float
     ) : UpdaterScreenState(version)
 
+    class Failed(
+        val failedReason: FailedReason
+    ) : UpdaterScreenState(null)
+
     object Rebooting : UpdaterScreenState(null)
 
     object CancelingUpdate : UpdaterScreenState(null)